move brew loading from install script to dotfile
Despite being coupled to zsh for now, there seems to be no other choice,
as the profile file will be overriden by chezmoi once the installation
completes, and any attempt to insert content into these files will cause
chezmoi to complain as it sees an uncommited diff.
By making `.zprofile` a template and enabling brew-loading only
if brew is installed we solve this problem.