always run configure-vpn script for sedg envs
It will always be run last, after all other files have been applied,
and it will not do anything if the tool has already been configured.
This is done to repace the "run once" behavior, as it caused
chezmoi to run the script before all files were applied,
leading to an erroneous state.