replace tool management with simple shell guards
Until now, special tools that had special configuration were marked
as installed in chezmoi's config file, but no tool other than `brew`
could be properly setup during installation, and had to be updated
manually, otherwise, its configuration won't be applied by chezmoi.

A better approach is to avoid managing them with chezmoi,
and instead guard the configs with checks for the existence of the tool,
a simple check that can be performed on any shell, on any machine.

Old machines that already had tool management as part of chezmoi's
config will be "migrated" once migrations are supported.