attempt to forcefully remove standalone chezmoi
At the end of the bootstrap process, when the standalone chezmoi is
replaced by a brew-managed one (if brew is installed), the replacement
is actually removal. This should be a forceful removal to avoid
false negatives that show as if the action has failed if the standalone
binary wasn't available to begin with (e.g. chezmoi was already
installed with some other tools).