fix how `ref` option is parsed in wrapper script
There are 2 ways to pass the `ref` option - Either separated
by a space or by an equal sign. The wrapper script
was only parsing the first case. This commit fixes
the parsing to handle both cases.

The implications of it were that the wrong ref was being
downloaded from the remote, causing the process to fail
when there are features that don't exist in the default branch.