Enhance macOS installation support and script improvements (#10)
* enhance install script to support mac

* fix real installer to properly support mac

Mac has some things that are done a bit differently
than Linux which we need to account for.

* try to install getopt via brew to be able to parse arguments

* properly unset variables in shell script before going out of scope

In POSIX shell scripts, it is a good practice to unset
variables before they go out of scope.
This helps to avoid potential issues with variable name
collisions and ensures that the script behaves as expected.

* verify bash version is at least 4.4, try installing newer if not

Installation would probably work properly only via brew,
which currently is used only for MacOS systems at this stage.
Given that this is mostly required for MacOS anyway, this is
fine for now.

* pass modified getopt environment to impl installer on macOS

We need a specialized version of the `getopt` program,
specifically, the GNU version of `getopt`, which the impl
script requires.

* replace all `hash` instances with `command -v` in install scripts

* fix brew checking command in installer-impl script

It was accidentally returning that brew was not installed when it was
for single-user installs (the most common case).

* add ability to invoke the install script locally

This allows running the entire "stack", both the wrapper script
and the implementation one locally, which is useful for testing.

* fix inconsistent system type used for MacOS in installer scripts

In some places the term `mac` was used, in others `darwin`.
To make it consistent, we use `darwin` everywhere as the
system type, and `mac` as the distribution name.

* fix download tool detection in wrapper script

Changing `hash` to `command -v` in the wrapper script to
had an implication that the path of the tool was printed if it
exists, causing the variable that captures the output to be
filled with much more then just the name of the tool as we
expected.

* tidy up system detection prints

* fix how `ref` option is parsed in wrapper script

There are 2 ways to pass the `ref` option - Either separated
by a space or by an equal sign. The wrapper script
was only parsing the first case. This commit fixes
the parsing to handle both cases.

The implications of it were that the wrong ref was being
downloaded from the remote, causing the process to fail
when there are features that don't exist in the default branch.

* fix infinite recursion in install-impl script when calling brew

We wrap the `brew` command in a function to handle brew
on multi-user systems, but inside it, for single-user systems,
we call `brew` again, which causes infinite recursion.
Instead we should prefix it with the `command` builtin to
avoid the function call.11

* setup brew for macOS as well in `.zshenv`

* fix `.zshrc` trying to load omz from hardcoded home path

On MacOS, the home location is different, and is not necessarily under
the `/home` directory (more like it's `/Users`).