Refactor Zsh configuration and switch to Sheldon (#11)
* remove omz and antigen from zshrc

* use `sheldon` as plugin manager

* track sheldon's plugin list in our dotfiles

* add support for `powerlevel10k` in zshrc

It's installed as a plugin using `sheldon`.

* git: configure default pager to be `diff-so-fancy`

* sheldon: remove unnecessary plugins

* move most of zshenv to zshrc

I once thought that zshenv was the best place to put all the
shell initialization code, but it turns out that zshrc is a better
place for most of it, as most of it is only needed in interactive
shells, and doing so in zshenv seriously slows down the startup
time of all shells, especially now that I'm using a deferring technique
to load the zshrc file in the background.

* use custom omz plugins that I have no alternative for

But instead of cloning omz itself, I use chezmoi's externals
to track specific files, each representing a plugin.

* fix keybindings by using oh-my-zsh's keybindings script

* remove unnecessary `update-omz` custom script

* fix issues with omz's `dotenv` plugin by manually setting cache dir

* sheldon: add plugin to replace `ls` and `tree` with `eza`

* add fzf-tab plugin, load all plugins at the end of the file

The reason to load them at the end is that some tools might
configure stuff that plugins respect, and won't override if set.
For example, fzf-tab will respect the original fzf configuration.

* improve fzf initialization when missing

* improve default fzf command

* include omz's git library module to make plugin complete

The plugin relies on some of the lib's functions to work properly