allow opt-in automatic install of system prerequisites
Most systems, be it an OS or a specific distro, have some prerequisites
that must exist before the installation is attempted. For example, these
are brew's prerequisites, assuming brew is used as an alternative
package manager later. We already have the funcionality to report
missing prerequisites and display hints on how to install them, now
we've added the ability to automatically try and install them (following
the same recipe as stated by the hints). It only works in
non-interactive mode for now.