fix tests to use default logger instead of mock
After adding many logging calls in latest commits, many tests broke as
they were using a mock logger, and with `moq`, each called method has to
be defined explicitly, otherwise the test fails. By replacing this mock
logger with the default logger, which is a noop logger, tests are
passing again and we don't have to customize the mock logger based on
what log-level it calls.