improve nested progress reports
Instead of trying to further indent different levels and somehow
maintain context with the parent level, an approach similar to Rust's
Cargo was implemented, where a progress updates its text, appending
deeper levels to the same line, until the topmost parent is complete,
which displays the item as completed.

Also, to allow for greater detail in some scenarios and "persist" the
actions that happened during a progress, somewhat resembling a log, a
new "Peristent Progress Logger" was introduced, for which
accomplishments can be logged as those in-between items. Again, much
like Cargo's behavior when compiling code, for example.