support interactive progress trackers
These are progress reporters that "pause" when an interactive operation
is about to be run, such as executing an external app that prints
instructions to stdout, and expects something from stdin. In our
context, this happens when creating GPG keys. Without this dedicated
support, the progress reporter, if interactive, would overlap with the
output from the other-interactive app, creating complete chaos in the
terminal.