make sure brew is available for mac early
For mac, prerequisites are expected to be installed via brew, but if
this installer is run on a fresh system that still hasn't installed
brew, it would fail. By installing brew earlier on macs, we ensure we
can handle prerequisites safetly for mac, using brew as its default
system package manager.