lower log level of package-managers version warning
Both package manager implementation we have, apt and brew, do not
support version constraints if those are passed next to a package to
install. Until now a warning was issues, but it's not really that big of
a deal, and should be made `debug` instead to avoid interfering with the
progressive installer.