fix race condition in progress logger tests
The tests tried to directly access the same buffer used by the
underlying `huh` library for showing the progress without any guards in
place, because `bytes.Buffer` isn't thread safe, and that caused Go's
race detector to report races. By implemeneting a safe buffer (and
writer), and reading from it safely in the tests themselves, the problem
is mitigated.