feature: remove support for multi-user systems (#13)
* remove support for multi-user systems

It never relaly worked in practice, so this just adds complexity and
noise at this stage. Now that our installer is fully compliant with the
depracted shell installer, we have no need for the mutli-user support.

* remove references to chezmoi's data multi-user entry

* add ability to apply dotfiles for custom branches

This is critical to get this PR passing CI, as our CI currently tries to
apply dotfiles from `main`, but they're incompatible with our installer
as they reference data that isn't there anymore. We need the CI to use
the current branch it's running on for full compatibility.