fix prerequisites not installing correctly
For advanced scenarios such as the development-tools group in DNF, our
implementation was losing critical information about the package, trying
to install it as a single-package rather than as a group. The missing
link was to use the package resolver which was built exactly for this
purpose, extracting as much info on the package as possible first, then
using it to properly install it.