feature: support distros using DNF as their package manager (#14)
* support special package manager types

Some package managers such as dnf and pacman support concepts other than
than package name itself, such as groups. These are now specified in the
yamls, and parsed correctly.

Would be used later in installer implementations.

* add support for installing dnf packages

It includes regular packages and the new group packages. Also completed
remaining package mapping for dnd, and marked comaptible distros that
use dnf as supported.

* ci: add e2e tests for fedora and centos

* fix prerequisites not installing correctly

For advanced scenarios such as the development-tools group in DNF, our
implementation was losing critical information about the package, trying
to install it as a single-package rather than as a group. The missing
link was to use the package resolver which was built exactly for this
purpose, extracting as much info on the package as possible first, then
using it to properly install it.

* use same keys for prerequisite packages

The resolver's job is to translate them to different names per-distro.

* add support for sub-package mappings per distro

"Development Tools" is a group package in DNF that has different
variants between Fedora and CentOS, named `development-tools` in Fedora
instead. This creates a complex situation where some packages are not
only mapped differently between different package managers, such as apt
and dnf, but also between distros on the same package manager.

By explicitly providing a mapping per-distro for such packages, we
ensure they'll work on specific distros, and error out if no mapping is
provided, thus not supporting anything implicitly.

* ci: test using latest fedora & centos

No need to pin a specific version for the latest test

* explicitly mark rocky and almalinux as unsupported

I've never used them so I don't want to take the chance.